+++ source/PyMaterialX/PyMaterialXRender/PyImageHandler.cpp
@@ -72,8 +72,14 @@ void bindPyImageHandler(py::module& mod)
         .def_static("create", &mx::ImageHandler::create)
         .def("addLoader", &mx::ImageHandler::addLoader)
         .def("saveImage", &mx::ImageHandler::saveImage,
-            py::arg("filePath"), py::arg("image"), py::arg("verticalFlip") = false)
-        .def("acquireImage", &mx::ImageHandler::acquireImage)
+            py::arg("filePath"), py::arg("image"), py::arg("verticalFlip") = false, py::arg("message") = (std::string*) nullptr)
+        .def("acquireImage", &mx::ImageHandler::acquireImage,
+            py::arg("filePath"), py::arg("generateMipMaps") = true, py::arg("fallbackColor") = (mx::Color4*) nullptr, py::arg("message") = (std::string*) nullptr)
+        .def("acquireImage", [](mx::ImageHandler& handler, const mx::FilePath& filePath, bool generateMipMaps, const mx::Color4* fallbackColor)
+            {
+                // Convert from v1.37.2 function signature.
+                return handler.acquireImage(filePath, generateMipMaps, fallbackColor);
+            })
         .def("bindImage", &mx::ImageHandler::bindImage)
         .def("unbindImage", &mx::ImageHandler::unbindImage)
         .def("setSearchPath", &mx::ImageHandler::setSearchPath)